backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]
python-jose
